get user data in one place
