+++ DM-APP-2024/Firebase/UserManager.js
@@ -1,5 +1,5 @@
 import {db} from "./firestore.js"
-import { collection, addDoc, getDocs } from "firebase/firestore"; 
+import { collection, addDoc, getDocs, doc, getDoc } from "firebase/firestore"; 
 import {auth} from "./AuthManager.js"
 
 let addUser = async (userData, uid) => {
@@ -7,9 +7,27 @@ let addUser = async (userData, uid) => {
         const docRef = await addDoc(collection(db, "Users"), {
             donorID: userData.donorID, email: userData.email, role: userData.role, team: userData.team, uid: uid, points: 0
         })
-        console.log("Document written with ID: ", docRef.id);
       } catch (e) {
         console.error("Error adding document: ", e);
       }
 }
-export {addUser}
+
+const getUserData = async () => {
+  try {
+    const currentUID = auth.currentUser.uid;
+    const docRef = doc(db, "Users", currentUID);
+    const docSnap = await getDoc(docRef);
+    if (docSnap.exists()) {
+      const data = docSnap.data();
+      return data;
+    } else {
+      console.log("Document does not exist");
+      return null;
+    }
+  } catch (error) {
+    console.error("Error fetching document data:", error);
+    throw error;
+  }
+};
+
+export { addUser, getUserData }
